airflow/requirements.txt
kafka-python==2.0.2
bs4
lxml